requirements.txt
# This file may be used to create a conda environment using:
# $ conda create --name <env> --file <this file>

python-telegram-bot==11.1.0
tensorflow==1.14.0
matplotlib==3.0.3
nltk==3.2.5
pillow==5.2.0
